Junction: add isNonTerminal flag
Traditionally Junctions were always splitting points, meaning that
Junctions could only be found on terminal points of lines.

Using a line merge strategy (like slope overlap resolution) this is no
longer true. The "isNonTerminal" flag is intended to be set on Junction
points that lie on the body of both intersecting lines.