Updated make binary function
Changed makeBinary function to also draw estimated width when Estimate
Width is ticked. When applying the plugin on a multi stack Composite
image, the binary image produced contains the same number of channels
and slices than the original image, each containing a binary image of
the detected lines. LUT of composite image is kept in binary image.